viddef_t changes, #1/3: Removed several int casts from earlier patches which are intended for silencing -Wsign-compare warnings.
git-svn-id: file:///svn/uhexen2/code/trunk@4580 eaf649e6-6c13-0410-b194-84221022858e

+++ engine/h2shared/draw.c
@@ -251,7 +251,7 @@ void Draw_Character (int x, int y, unsigned int num)
 	if (y <= -8)
 		return;			// totally off screen
 
-	if (y > (int)vid.height - 8 || x < 0 || x > (int)vid.width - 8)
+	if (y > vid.height - 8 || x < 0 || x > vid.width - 8)
 		return;
 
 	row = num >> 5;
@@ -467,7 +467,7 @@ void Draw_SmallCharacter (int x, int y, int num)
 		num -= 32;
 	}
 
-	if (y >= (int)vid.height)
+	if (y >= vid.height)
 	{ // Totally off screen
 		return;
 	}
@@ -479,7 +479,7 @@ void Draw_SmallCharacter (int x, int y, int num)
 	}
 #endif
 
-	if (y + 5 > (int)vid.height)
+	if (y + 5 > vid.height)
 	{
 		height = vid.height - y;
 	}
@@ -648,10 +648,8 @@ void Draw_Pic (int x, int y, qpic_t *pic)
 	byte		*source;
 	int		v, u;
 
-	if ((x < 0) ||
-		(x + pic->width > (int)vid.width) ||
-		(y < 0) ||
-		(y + pic->height > (int)vid.height))
+	if ((x < 0) || (x + pic->width > vid.width) ||
+	    (y < 0) || (y + pic->height > vid.height))
 	{
 		Sys_Error ("%s: bad coordinates", __thisfunc__);
 	}
@@ -705,12 +703,12 @@ void Draw_PicCropped (int x, int y, qpic_t *pic)
 		Sys_Error("%s: bad coordinates", __thisfunc__);
 	}
 
-	if (y >= (int)vid.height || y+pic->height < 0)
+	if (y >= vid.height || y+pic->height < 0)
 	{ // Totally off screen
 		return;
 	}
 
-	if (y+pic->height > (int)vid.height)
+	if (y+pic->height > vid.height)
 	{
 		height = vid.height-y;
 	}
@@ -792,8 +790,8 @@ void Draw_TransPic (int x, int y, qpic_t *pic)
 	byte		*source, tbyte;
 	int		v, u;
 
-	if (x < 0 || (unsigned int)(x + pic->width) > vid.width || y < 0 ||
-			(unsigned int)(y + pic->height) > vid.height)
+	if (x < 0 || (x + pic->width) > vid.width ||
+	    y < 0 || (y + pic->height) > vid.height)
 	{
 		Sys_Error("%s: bad coordinates", __thisfunc__);
 	}
@@ -882,17 +880,17 @@ void Draw_SubPicCropped (int x, int y, int h, qpic_t *pic)
 	byte		*source;
 	int		v, u, height;
 
-	if ((x < 0) || (x+pic->width > (int)vid.width))
+	if ((x < 0) || (x+pic->width > vid.width))
 	{
 		Sys_Error("%s: bad coordinates", __thisfunc__);
 	}
 
-	if (y >= (int)vid.height || y+h < 0)
+	if (y >= vid.height || y+h < 0)
 	{ // Totally off screen
 		return;
 	}
 
-	if (y+pic->height > (int)vid.height)
+	if (y+pic->height > vid.height)
 	{
 		height = vid.height-y;
 	}
@@ -982,17 +980,17 @@ void Draw_TransPicCropped (int x, int y, qpic_t *pic)
 	byte		*source, tbyte;
 	int		v, u, height;
 
-	if ((x < 0) || (x+pic->width > (int)vid.width))
+	if ((x < 0) || (x+pic->width > vid.width))
 	{
 		Sys_Error("%s: bad coordinates", __thisfunc__);
 	}
 
-	if (y >= (int)vid.height || y+pic->height < 0)
+	if (y >= vid.height || y+pic->height < 0)
 	{ // Totally off screen
 		return;
 	}
 
-	if (y+pic->height > (int)vid.height)
+	if (y+pic->height > vid.height)
 	{
 		height = vid.height-y;
 	}
@@ -1164,10 +1162,8 @@ void Draw_SubPic (int x, int y, qpic_t *pic, int srcx, int srcy, int width, int 
 	byte		*source;
 	int		v, u;
 
-	if ((x < 0) ||
-		(x + width > (int)vid.width) ||
-		(y < 0) ||
-		(y + height > (int)vid.height))
+	if ((x < 0) || (x + width > vid.width) ||
+	    (y < 0) || (y + height > vid.height))
 	{
 		Sys_Error ("%s: bad coordinates", __thisfunc__);
 	}
@@ -1211,8 +1207,8 @@ void Draw_TransPicTranslate (int x, int y, qpic_t *pic, byte *translation, int p
 	byte		*source, tbyte;
 	int		v, u;
 
-	if (x < 0 || (unsigned int)(x + pic->width) > vid.width || y < 0 ||
-			(unsigned int)(y + pic->height) > vid.height)
+	if (x < 0 || (x + pic->width) > vid.width ||
+	    y < 0 || (y + pic->height) > vid.height)
 	{
 		Sys_Error ("%s: bad coordinates", __thisfunc__);
 	}
@@ -1638,8 +1634,8 @@ void Draw_Fill (int x, int y, int w, int h, int c)
 {
 	int		u, v;
 
-	if (x < 0 || x + w > (int)vid.width ||
-		y < 0 || y + h > (int)vid.height)
+	if (x < 0 || x + w > vid.width ||
+	    y < 0 || y + h > vid.height)
 	{
 		Con_Printf("Bad Draw_Fill(%d, %d, %d, %d, %c)\n", x, y, w, h, c);
 		return;
@@ -1713,7 +1709,7 @@ void Draw_FadeScreen (void)
 	for (x = 0; x < 2048; x++)
 		temp[x] = (164 + rand() % 6) * 256;
 
-	for (y = 0; y < (int)vid.height; y++)
+	for (y = 0; y < vid.height; y++)
 	{
 		pbuf = (byte *)(vid.buffer + vid.rowbytes*y);
 		pos = &temp[rand() % 256];
@@ -1725,7 +1721,7 @@ void Draw_FadeScreen (void)
 			VID_LockBuffer ();
 		}
 
-		for (x = 0; x < (int)vid.width; x++, pbuf++)
+		for (x = 0; x < vid.width; x++, pbuf++)
 		{
 //			if ((x & 3) != t)
 //				pbuf[x] = 0;
